+++ lib/jay_api/elasticsearch/stats.rb
@@ -2,6 +2,8 @@
 
 require_relative 'stats/index'
 require_relative 'stats/indices'
+require_relative 'stats/node'
+require_relative 'stats/nodes'
 
 module JayAPI
   module Elasticsearch
@@ -21,18 +23,36 @@ module JayAPI
       #   to the Statistics API endpoint fails.
       def indices
         # DO NOT MEMOIZE! Leave it to the caller.
-        ::JayAPI::Elasticsearch::Stats::Indices.new(response['indices'])
+        ::JayAPI::Elasticsearch::Stats::Indices.new(indices_stats['indices'])
+      end
+
+      # @return [JayAPI::Elasticsearch::Stats::Nodes] Information about the
+      #   nodes that make up the Elasticsearch cluster.
+      # @raise [Elasticsearch::Transport::Transport::ServerError] If the request
+      #   to the Statistics API endpoint fails.
+      def nodes
+        # DO NOT MEMOIZE! Leave it to the caller.
+        ::JayAPI::Elasticsearch::Stats::Nodes.new(nodes_stats['nodes'])
       end
 
       private
 
-      # @return [Hash] The Hash with the statistics returned by the
+      # @return [Hash] The Hash with the index-related statistics returned by
+      #   the Elasticsearch cluster.
+      # @raise [Elasticsearch::Transport::Transport::ServerError] If the
+      #   request fails.
+      def indices_stats
+        # DO NOT MEMOIZE! Leave it to the caller.
+        transport_client.indices.stats
+      end
+
+      # @return [Hash] The Hash with the node-related statistics returned by the
       #   Elasticsearch cluster.
       # @raise [Elasticsearch::Transport::Transport::ServerError] If the
       #   request fails.
-      def response
+      def nodes_stats
         # DO NOT MEMOIZE! Leave it to the caller.
-        transport_client.indices.stats
+        transport_client.nodes.stats
       end
     end
   end